refactor: extract GitHub API and Backport classes
To improve the flexibility, testability and maintainability, this commit
extracts the Github api into its own interface and class as well as
extracts the Backport code into its own class.

This allowed for a cleanup of the backport.test.ts file as well as
easier to read code in the main.ts, backport.ts and github.ts files.

Lastly, this will allow for easier fixups of types when updating to
newer versions of the github dependencies.