feat: add relates to issue links
Uses the events of the pull request (a pull request is also an issue,
and thus has events), to determine which issues were linked to it at any
time. Since only the `connected` events are used, it can be the case
that an issue was unlinked or `unconnected` from the pull request.
Unlinked issues are not taken into account, so they will be used as if
they were still linked.