fix(backport): checkout detached target from origin
By using DETACHED mode when creating the worktree, we make sure not to
conflict with any other checkout branches (a branch is only allowed to
be checked out once in all worktrees).

By using origin, we make sure that the correct target is used to
checkout. Potentially the user already had target checkout at an earlier
commit and has not yet brought it up-to-date.