fix(workflow): only run on merged pull requests
The pull_request [closed] event contains both merged and unmerged pull
requests. Since the bot now allows to trigger on comment, it has a build
in warning against being triggered for a unmerged pull request, but this
warning is also given when just closing a pull request.

This commit fixes this by further restricting when to trigger the action