feat: add pull description config option
To allow users to configure the description used in the pull requests
created by this action. The default stays the same.

In order to stay the same, and add something meaningful to the
descriptions. This also allows the use of placeholders to define
variable values.

For now, this is just a simple replacement of exact matches, but in the
future we could consider escaping these placeholders if there is a need
for it.