feat(utils): add get mentioned issue refs from body
This method was hard to build due to the complications of working with
regex. However, because the entire method can work on a string and does
not require any other dependencies, it made sense to expose it as a
single exported function in a utils file. This made testing a lot
easier, because it does not require us to do any additional setup.

The tests are structured in a way to test the most common cases and the
helper method is created to allow the test code to focus on test cases.

In ES2020, string.matchAll is added which allows capturing groups of all
matches. Using this we would not have to execute the url pattern on each
individual matched url again to find the capture groups (which we use to
construct the issue reference). Sadly, ES2020 is not available in nodejs
12 which is used by default for github actions.