fix(exec): split script into individual commands
In order to split the script into individually executable commands, a
new git.ts file is introduced. Since these changes also required changes
to the acceptance tests (the script could no longer be executed as a
whole), this commit also replaces @actions/exec and child_process with
execa. This library makes executation of commands a bit easier and can
be used inside a github action as well as locally in tests.

While making changes, I was tempted into replacing the use of git
worktree for cherry-picking. Error handling has also changed a bit,
because in the script this was handled by `-x` flag, but must be done
explicitly in typescript.

Lastly, the acceptance test has had some minor changes, to improve
analysing failing tests.