fix(backport): fetch pull head and targets
By fetching the pull head, we make sure that we have all commit shas
available, even if the branch itself is already deleted from github.

Ofcourse, we only want to fetch the head once, since it doesn't change for
each of the different targets. And if none of the labels match, then we
don't even have to fetch.

Potentially this allows us to remove the `fetch-depth: 0` input from the
checkout action. To achieve that goal, we'll need to also fetch each of
the targets individually.