fix: remove auto request reviewers
The auto request reviewers feature never really worked well. The problem
is that it tries to copy the current open review requests, but when a PR
is merged, most of those reviews have already been completed. When
completed, they are no longer in the review request list.

The feature became problematic for users when using a bot user (PAT)
that does not have admin rights to the repository. Those rights are
necessary to request reviewers in some repos.

Multiple solutions were possible, but as this feature doesn't behave
as wanted, and might even be unexpected to some users (breaking the
principle of least astonishment), I believe it should simply be removed.
We can always re-introduce it with a configuration option at a later
time.