fix: use direct commit shas in cherry-pick
It seemed that ranges behave strangely when cherry-picking in a shallow
clone, but I was mistaken. The commit range `sha1..sha2` is a shorthand
for `^sha1 sha2`, which means all commit reachable by sha2 but not
reachable by sha1. That means that sha1 is excluded.

So to make the cherry-pick work correctly with the shorthand (..)
notation, we'd need to use something like: `git cherry-pick
<firstCommitSha>^..<lastCommitSha>`, i.e. take all commits reachable
from <lastCommitSha> but not reachable from the first parent of
<firstCommitSha>. This would include the commit referenced by
<firstCommitSha> as well.

Instead of using ranges, we can pass all specific commits directly to
the cherry-pick command. This isn't easier, but makes the command more
explicit. We can always choose to rewrite this back to using the
shorthand (..) notation.