perf(git): only fetch pull head and target
The base ref was necessary to determine which commits exist on the pull
request. The common ancestor of base ref and pull head could be found
using git merge-base. However, many commits may exist between the base
ref and this common ancestor. In some cases this could lead to a slow
fetch of the base ref, because it requires a deep fetch.

Instead of determining the commits on the pull request using git only,
we can simply ask github about the commits on the pull request. Github
can provide the number of commits and there is an API to retrieve data
about the commits in a paged way. Sadly this is limited to 250 commits,
but that should cover most cases. We can improve upon this in the
future.

Using the number of commits on the pull request we can fetch exactly
those commits needed for cherry-picking.

We can then retrieve all the commits on the PR to find the first and
last commit shas.

Finally we can cherry-pick the entire range of commits:
`git cherry-pick firstCommitSha..lastCommitSha`

Note that this completely removes the old backport script method, which
most tests depend on. It means we have remove/disable most tests from
this repo.

In the meantime, backport-action-test provides a way to validate the
correct behavior of this action.