docs(README): improve usage example
The names no longer align with the action's description, which has moved
away from PR to pull request.

It also even called the job `build` which is just misleading.

Lastly, the permissions can be pulled up to the entire workflow, which
doesn't change the permissions for any of the steps, but improves the
readability of the workflow.

These changes are also applied to the 'trigger using a comment' example.