dist: rename action for marketplace
The name `backport-action` was already in use. Let's see if `Backport`
would be allowed.

Looking at the descriptions of the actions in the Marketplace, there's
not much to set it apart from the others. So, I've chosen to change the
description to highlight one main feature.

It might be useful to reflect the features in the README better.