fix(backport): comment error when target not found
When the target cannot be found on remote, then the git fetch fails
with: (example for `maintenance/8.1`)
```
fatal: couldn't find remote ref maintenance/8.1
'git fetch origin maintenance/8.1' failed with exit code 128
```

We can improve this by throwing a special error and handling it
explicitly by commenting on the original pull request about this
situation. We can also help the user figure out how to resolve the
problem by making it clear that a branch with that name should exist.