fix(github): use string array
Not sure if this would actually cause any issues, but `[string]` is
different from `string[]`. The latter is a string array, the first is an
array consisting of 1 string element. This should've prevented us from
passing an empty array and passing an array with multiple elements, but
I guess this doesn't work too well in TS. It's also not what we want.