ci(release): add release workflow
This workflow can be used to automatically release new versions of the
backport-action.

It sets the version in the package.json, and tags this commit.

It also creates (or moves) tags for the major and minor version.

It can also dry-run, meaning that it does not push any commits or tags.
The created commits are always logged.