docs(README): use pull_request_target
For contributions from forked repositories, pull_request_target is
needed so that GITHUB_TOKEN has write access to the repo. That is needed
for the action to push the commits it cherry-picked.

If pull_request_target is not allowed to be used, the action can also be
run with pull_request by checking out the repo with a PAT that has write
access. This is better for security as this does not expose the repo's
secrets to the action.

Please note that this action does not read any secrets anyways, but it
could be a consideration for some projects with strict security
policies.