ci(release): add release next preview workflow
This workflow runs on every commit to main and moves the next preview
along. In the future, we could also run it on the branches of previously
released major versions.

The workflow is automatically updated on new releases so we don't forgot
to update the major part of the version (v1.next-preview,
v2.next-preview).