ci(ci): publish artifacts on main branches
We were unable to eat our own dogfood, because the artifacts always had
to be committed along with the source code changes in a pull request.
This would lead to conflicts when backporting the changes of the pull
request to an older release branch. By publishing the artifacts in a
separate step of the workflow, we can avoid that the commit with the
artifacts is part of the pull request.

The artifacts should only be published on pushes to the `main` branch
and on branches starting with `release` (which will be the naming scheme
for long living release branches in this repo).

Packaging the artifacts should always be done, to make sure a pull
request's changes don't break packaging the artifacts.