ci(ci): checkout commit and push using pat
GHA have a rule that workflows cannot be triggered on events caused by
the GITHUB_TOKEN. So, the commit pushed by the CI (publishing the
artifacts) would not trigger the release next preview workflow. But, it
should, because the next preview tag should point to the commit with the
artifacts.

To circumvent this, we can checkout the repo in the workflow with a PAT.
This PAT will automatically be used when committing and pushing the
packaged artifacts.

In addition to this commit, the repo now needs to specify a secret:
`BACKPORT_ACTION_PAT`. There are no workflows run on pull_request, so
this should be safe (secrets are not exposed to pull_request_target),
and contributors can't push directly to main or release branches.