refactor: extract findTargetBranches
If we want to add an input that specifies target branches, then the
structure of the code needs to be slightly different. Previously, the
code would loop over each label and determine if a backport should
happen for that label. Now, we want to loop over all target branches.

This means that we have to find the target branches first, and then loop
over them.

This refactoring extracts this to a separate function, so it can be
easily called and tested.

It also determines the target branches before fetching all the commits
from the pull request. This allows us to remove the check whether any of
the labels matches the label pattern, because this allowed us to avoid
unnecessarily fetching the commits if there were no targets found. We
can now simply stop when there are no target branches found.