docs: order inputs alphabetically
As we continue to expand the number of available inputs, it makes sense
to order the inputs alphabetically. This allows users to more easily
find the documentation for the input they're looking for.

To help out with top-to-bottom readabability, we can use links in the
descriptions. For example, in the how it works section we can link
directly to the input's heading.