docs: remove on pull_request note
This note contained incorrect information.

It is true that a workflow that runs in the `pull_request` context for a
pull request from a forked repository does not have write access. It is
also true that a PAT can be used to specify permissions. But, a PAT
should be stored as a secret and secrets are not available to workflows
running in the `pull_request` context for pull requests from a forked
repository. So the proposed solution cannot work.

We can simply remove the proposed alternative, since it is fine to run
this workflow in the `pull_request_target`.