refactor: make headref available to find target branches
If we want to exclude the headref from the possible target branches, we
need to provide it as a parameter to the findTargetBranches function.

I'm not so happy that this affects all previously written unit tests for
this function, but it is acceptable because headref has now become
important for this function.