docs: exclude headref from selected target branches
The headref is excluded automatically. Since this is an explicit
difference in behavior, it should be clearly documented to avoid
surprises for users.

As far as I know, there is no use case to backport a pull request to
its headref. The cherry-pick would be interrupted because there are no
changes which would result in empty commits. Typically, the git user
get's to decide to skip these commits, or to abort the cherry-pick. This
means that would not be any new commits, and we cannot open an empty
pull request. So all of this doesn't work. The only thing left to do, is
exclude the headref from the determined target branches.