ci: publish artifacts on release branches
The CI was supposed to publish artifacts on the main branch and on
release branches. However, I misread the documentation for the
`contains(search, item)` expression. I expected that search is the value
to search for, while item is the thing to search in. It is actually the
other way around: search is the thing to search in, and item is what is
searched for.

In addition, contains might not be explicit enough. So for safeguarding
it may be better to use `startsWith(searchString, searchValue)`. This
function 'Returns true when searchString starts with searchValue'.

This problem should now be resolved.

(cherry picked from commit dceff6b56537f2ec8bd826a42aab5725b0ecbb68)