ci: run on pull_request
pull_request_target is run in the context of the pull request's base,
which makes sense because secrets are made available. But, it meant that
the checked out code that was tested was also just that from the base,
and not that of the pull request.

To avoid this, I could checkout the code from the pull request's head,
but that is dangerous as it may expose secrets to malicious actors.

Instead, its better to just revert to running on pull_request instead.