ci: dont use secret for checkout
The secret was used to checkout so the pushed commit (Publish Artifacts)
would be able to trigger the release next preview workflow.

A secret was needed because the standard github.token cannot produce
events that trigger other workflows.

However, secrets cannot be used in pull requests coming from forks,
which was causing issues with contributions and pull requests from
dependabot.

It's also not wise to run CI with pull_request_target while checking out
the pull request's code because this may expose secrets to malicious
actors.

In the end, I see no other solution as to inline the additional workflow
as another step.