ci: publish artifacts after merging
So after all the recent changes, I forgot to adjust the references used
to decide whether the artifacts are published and whether the
release-next tag is moved.

These refs should be those of the pull_request base (like main and
release). Additionally, I just realized that I only have to check these
for the move next preview tag step. The artifacts can always be build
when the pull request is merged.

However, this made me realize that the artifacts should be published by
commit (and pushing) to the pull request's base ref as well. Not to the
pull request's head.