refactor: inject execa and git
Inversion of control makes it easier to test backport.ts and git.ts.

I had to add a bit of a strange type for execa, using the import()
function. This is because execa does not provide the types in an easily
importable way. If you know of a better way to do this, please let me
know.

The idea of this type is to take the type of the import of execa, which
is an object that contains all the functions of the execa package.

```
type Execa = (typeof import("execa"))...
```

Then, we take the `execa` function from this object using an indexed
access type:

```
type Execa = ...["execa"]
```