feat: filter out merge commits
Merge commits cannot be cherry-picked unless a mainline is specified. In
other to add the skip merge commit behavior as an input, we need to have
a way to remove the merge commits from the shas to cherry pick.

We can use git rev-list --merged to list the merge commits in a commit
range. Since we don't have a range but an array of commits, we can build
the shorthand for the range as the `first^..last`.