feat: add merge_commits input
Used to set the commits.merge_commits config setting which either fails
the backport when the action encounters merge commits on the merged pull
request, or skips any merge commits when cherry-picking.

Misconfiguration fails the workflow run because this requires attention
from the workflow owner. The fail setting fails the backport as
described in the README, i.e. the action reports on its success or
failure through a comment on the merged pull request.

Perhaps in a future version I'll support failing the workflow run on
failing a backport, but for now I'll keep that out of scope.