docs: specify false as default for new inputs
The inputs expect `true` to enable the copying behavior. It wasn't
really clear from the action.yml nor from the readme that these are
boolean flags. By specifying the default as `false` this becomes clear.

Ultimately, an empty input would still disable the copying behavior, but
that's less relevant than the knowledge that this is a boolean flag.

This change doesn't affect the usage or implementation.