fix: suggest to cherry-pick specific commits
The manual instructions still suggested to determine the commits to
cherry-pick by determining the common ancestor (merge-base) of the head
and base refs, first. However, this way has been abandoned for quite
some time now. This especially became a problem now that merge commits
can be skipped, meaning that the suggestion no longer produced a
reliable result.

Instead, we can simply suggest the same as the action does. It knows
exactly which commits to cherry-pick.