docs: only start backport by comment starting with
the `contains`  expression can lead to unexpected action boot-loops when the error itself contains `/backport`.

This can be the case when a branch is includes the name "backport".

for example in a comment like this
```
error message .. in branch origin/backport-form-xx-to-xx
```

Using `startsWith` to reduce the likeliness of this issue.
https://docs.github.com/en/actions/learn-github-actions/expressions#startswith