feat: introduce experimental detect_merge_method
The experimental input allows more flexibility when adding new
experimental features, as well as the ability to remove them again
without breaking things.

For unknown specified experimental properties, the action can log
warnings.

The input property also clarifies to users that this concerns an
experimental feature, meaning that bugs may be present or that the
feature breaks user space when enabled.

The spread operator can be used quite simply to set some defaults and
overwrite them.