perf: only request merge_commit_sha once
The merge commit sha was requested in the mergeStrategy method to
determine the merge strategy, and when the strategy is squashed, it
would retrieve the same sha again.

We can avoid one of these calls to the GH API by requesting this value
early. We need anyways to continue determining the method strategy.

This way, we can also reuse the value for the rebased strategy when we
implement a different picking approach there.

Note that the merge_commit_sha can be either a string or null, but when
we've determined that it's a squashed merge, we know that the
merge_commit_sha is set. So we can use the ! (non-null assertion
operator) to instruct typescript that it's really just a string at this
point.