feat: detected rebase cherry-picks rebased commits
I believe this functionality makes most sense to users. If a rebase is
detected they expect the rebased commits to be cherry-picked rather than
the commits from the pull request.

To find these commits, we can use git log over a range of commits.
Starting from the merge_commit_sha (which is the commit that the base
branch was updated to) we want to cherry-pick n number of commits.

Let's say there are 3 commits on the base branch after merging the pull
request:
* abc
* def
* ghi
* ..

We can find these three commits by traversing all commits from abc
excluding any commits reachable from the parent of ghi. In other words
we can do `git log ^ghi^ abc`. Note that we don't know the sha of ghi
nor its parent, but we know how to reach it. We can find the parent of
ghi as abc~3. This means we can do: `git log ^abc~3 abc` for which the
shorthand notation is `git log abc~3..abc`. Note that the number 3 can
be replaced by the number of commits on the pull request.

To do this, we must ensure that these commits are known locally. So we
must fetch these commits (and +1 in case we're in a shallowly cloned
repo).