fix: store fetched refs indefinitely
When a commit is fetched directly by sha, it is only tracked by
FETCH_HEAD. However, FETCH_HEAD gets overwritten in subsequent calls.

To ensure we store the fetched commit, we must specify a refspec in the
form of `+src:dst`. If <dst> is not an empty string, an attempt is made
to update the local ref that matches it.

To find a unique ref to use as dst we can use the src sha.

It may be that this requires --force but I think we should try without
first.