fix: check commits to cherry pick for merges
In order to skip merge commits, we check for merge commits among the
commits to cherry pick. Now that those have changed due to
squash/rebase merge methods, we should check the correct commits.

Note, it's a bit useless to check these commits for merge commits, but
we can optimize it later. I'd rather be strict about what we're
checking.

This does still fix a case where a merge commit exists on the pull
request that was rebased into a non-merge commit.