breaking: change default cherry-picked commits
The experimental input is replaced with a new non-experimental input
`cherry_picking`. This changes the behavior, as the new input has a
different default behavior.

From now on commits are cherry-picked based on the merge method by
default. It's still possible to cherry-pick the commits from the pull
request itself, by changing the cherry_picking input.