fix: fetch and checkout draft pr branch directly
When the pull request is created in draft mode because of conflicts, the
suggestion did not fetch the branch of the opened draft pull request.
This meant that the instructions fail when switching, and the commits
are cherry-picked on the wrong branch.

Instead, we can fetch the branchname immediately. We no longer need to
fetch target either. And while we're at it, we can also simply checkout
the branch immediately, so we don't have to switch to it separately.