fix: ignore existing branches
Don't throw an error when the backport branch already exists from a
previous run of the action.

We intentionally don't `continue` to the next target, because a previous
run might have created the branch successfully, but failed to create a
PR. Thus we just keep going and try to create the PR as well.