fix: log existing branch recovery
It's useful to explain what happens to the readers of the logs. In the
case where the push of the backport branch failed, we attempt to recover
a previous run by verifying that a branch with the same name already
exists on the remote. If so, we can simply create a PR for it.

It's possible that the remote branch does not contain the same contents,
and we currently don't verify this. However, malicious actors would need
write access to the repo to push such a branch anyways. So, this is not
considered to be dangerous.