feat: add experimental backport dashboard
This adds a new input dashboard_enabled that allows the action to create
or update a single GH issue for tracking the status of backport pull
requests, i.e. the backport dashboard.

The dashboard can be found by title, or created newly if it doesn't
exist yet.

The issue is used as a database, so it's contents are parsed and can be
updated again. This means that users can manually track backport pull
requests that were created before the introduction of this feature
simply by editing the dashboard issue.

Before releasing this, we'll need to consider the security implications
of parsing user-editable data.