fix: closed backports are no longer pending
The goal is that we track pending backports, because we want to ensure
that each original PR that required backporting is also actually fully
backported eventually, i.e. merged.

However, any closed backport pull request is likely closed
intentionally. It can be considered as the closer stating: this backport
is no longer required. Of course, it could be that a manual backport is
still performed separately, but that's not a PR created by backport
action, so it can't track it.

For the purpose of tracking pending backport pull requests created by
the action, I believe we can simply consider closed backports as no
longer active.