fix: use the oldest dashboard issue only
We'll need to somehow avoid that issues accidentally get overwritten.
For example, when someone creates a new issue with the exact title
'Backport Dashboard' then it could be found by the action and updated,
overwriting the original contents.

Of course, GitHub offers a version history for each issue, but we can
provide some ways to avoid this issue.

One way to do so would be to allow the workflow owner to specify the
dashboard issue by id. However, we don't have that option yet.

However, we need a good default behavior. We can simply make the choice
to select only the oldest open Backport Dashboard issue. It's up to the
maintainer to ensure that there is no older issue with the same title.
This way, newer issue will never conflict.

The only cases where this can fail, is that an older issue with the same
title is reopened, or an older open issue is renamed. To avoid these,
we'll need to support an input that allows specifying the issue id
directly. We can do that later.